formatted the code

+++ contracts/Item.sol
@@ -1,27 +1,36 @@
-// SPDX-License-Identifier: GPL-3.0 
+// SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.0;
 
 import "./ItemManager.sol";
- contract Item {
-    uint public priceInWei;
-    uint public paidWei;
-    uint public index;
-    
+
+contract Item {
+    uint256 public priceInWei;
+    uint256 public paidWei;
+    uint256 public index;
+
     //Created parentcontract of type ItemManager
     ItemManager parentContract;
-    
+
     //Calling the constructor
-    constructor(ItemManager _parentContract, uint _priceInWei, uint _index)  {
-        priceInWei =_priceInWei;
-        index= _index;
-        parentContract= _parentContract;
-         }
-     receive() external payable {
-         require(msg.value==priceInWei, "We don't support partial payments");
-         require(paidWei==0, "Item is already paid !");
-         paidWei+=msg.value;
-         (bool success,)= address(parentContract).call{value:msg.value}(abi.encodeWithSignature("triggerPayment(uint256)", index));
-         require(success, "Delivery did not work");
-     }
-     fallback () external {}
+    constructor(
+        ItemManager _parentContract,
+        uint256 _priceInWei,
+        uint256 _index
+    ) {
+        priceInWei = _priceInWei;
+        index = _index;
+        parentContract = _parentContract;
+    }
+
+    receive() external payable {
+        require(msg.value == priceInWei, "We don't support partial payments");
+        require(paidWei == 0, "Item is already paid !");
+        paidWei += msg.value;
+        (bool success, ) = address(parentContract).call{value: msg.value}(
+            abi.encodeWithSignature("triggerPayment(uint256)", index)
+        );
+        require(success, "Delivery did not work");
+    }
+
+    fallback() external {}
 }
